backend/pyproject.toml
[project]
name = "ai-cv-personalizer-backend"
version = "0.1.0"
description = "AI CV Personalizer Backend with CrewAI integration"
authors = [{ name = "KPATCHA ESSOHANAM JOSUE", email = "josuekpatcha1@gmail.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "crewai[tools]>=0.150.0,<1.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
serve = "main:run_server"
crew = "crew_main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
